Change type of `target_score` parameter of `NonceProvider::nonce` from `f64` to `u32` to better match TIP32 (#1503)
* Change type of `target_score` parameter of `NonceProvider::nonce` from `f64` to `u32` to better match TIP32

* Changelog

+++ bee-pow/src/providers/mod.rs
@@ -33,5 +33,5 @@ pub trait NonceProvider: Sized {
     }
 
     /// Provides a nonce given bytes and a target score.
-    fn nonce(&self, bytes: &[u8], target_score: f64) -> Result<u64, Self::Error>;
+    fn nonce(&self, bytes: &[u8], target_score: u32) -> Result<u64, Self::Error>;
 }
